# src/main/swagger/anime-recommender/api.yaml
openapi: 3.0.1
info:
  title: Anime Recommender Service
  description: 'This is a server for the Anime Recommender Application.  '
  termsOfService: http://swagger.io/terms/
  contact:
    email: estoeban@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
  - url: https://estoeban.io/v1
  - url: http://estoeban.io/v1
tags:
  - name: Recommendation
    description: Recommendations for user
    externalDocs:
      description: Find out more
      url: http://swagger.io
  - name: user
    description: Operations about user
paths:
  /recommendations:
    post:
      tags:
        - Recommendation
      summary: Fetch Recommendations for user
      operationId: getRecommendations
      requestBody:
        description: Reqest to fetch recommendations based on a root Anime
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecommendationRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/RecommendationRequest'
        required: true
      responses:
        200:
          description: OK
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/RecommendationResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationResponse'
        400:
          description: Bad Request
          content: {}
        404:
          description: Not Found
          content: {}
      x-codegen-request-body-name: body
components:
  schemas:
    RecommendationRequest:
      type: object
      properties:
        animeId:
          type: string
        degreesOfSeparation:
          type: integer
          format: int32
    AnimeRecommendations:
      type: array
      items:
        $ref: '#/components/schemas/Anime'
    Anime:
      type: object
      properties:
        animeId:
          type: string
        description:
          type: string
    RecommendationResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        recommendations:
          $ref: '#/components/schemas/AnimeRecommendations'